pyproject.toml
[project]
name = "collectorinfra"
version = "0.1.0"
description = "인프라 데이터 조회 에이전트 - 자연어로 인프라 DB를 질의하고 문서를 생성하는 AI 에이전트"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # 에이전트 프레임워크
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",

    # MCP 클라이언트 (DBHub 모드)
    "mcp",

    # DB 직접 연결 (direct 모드)
    "asyncpg>=0.29.0",

    # 체크포인트
    "langgraph-checkpoint-sqlite",

    # API 서버
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",

    # 유틸리티
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "sqlparse>=0.5.0",
    "structlog>=24.0.0",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "python-multipart>=0.0.9",
    "redis[hiredis]>=5.0.0",
]

[project.optional-dependencies]
document = [
    "openpyxl>=3.1.0",
    "python-docx>=1.0.0",
]
postgres-checkpoint = [
    "langgraph-checkpoint-postgres",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "ruff>=0.4.0",
    "mypy>=1.10",
]

[project.scripts]
collectorinfra = "src.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
